Move some of relevant code to ReflectionSupport.kt
